+++ src/config.ts
@@ -1,0 +1,14 @@
+export const alphabet =
+  'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_';
+
+export const timestampBitSize = 8;
+export const processIdBitSize = 6;
+export const serviceIdBitSize = 2;
+export const entityTypeIdBitSize = 2;
+export const counterBitSize = 6;
+
+export const timestampPositionStart = 0;
+export const processIdPositionStart = 8;
+export const serviceIdPositionStart = 14;
+export const entityTypeIdPositionStart = 16;
+export const counterPositionStart = 18;