+++ src/constants.ts
@@ -1,0 +1,12 @@
+export const radix16 = 16;
+
+export const second = 1000;
+
+// 255 - 8-bit - Tiny Int
+export const uTinyIntMax = 0xff;
+
+// 16777215 - 24-bit - Medium Int
+export const uMedIntMax = 0xffffff;
+
+// 4294967295 - 32-bit - Int
+export const uIntMax = 0xffffffff;