+++ src/encoder.ts
@@ -1,0 +1,8 @@
+import intEncoder = require('int-encoder');
+import { alphabet } from './config';
+import { radix16 } from './constants';
+
+intEncoder.alphabet = alphabet;
+
+export const decode = (arg: string): string => intEncoder.decode(arg, radix16);
+export const encode = (arg: string): string => intEncoder.encode(arg, radix16);